gsockaddr.h: remove includes that are not strictly needed for the interface
This patch removes a number of headers from gsockaddr.h, which is a pretty
widespread dependency in the codebase, due to it being included by
logmsg.h.

Albeit gsockaddr itself compiled cleanly with this change, a number of
depended modules relied on these headers to be present, causing a trickling
effect of functions/macros being undefined. These were also
fixed up in this patch.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>